Add files via upload
Tic-Tac-Toe is a two-player game in which the players fill up nine empty rectangles in a table with either an X or an O when it is their turn. Once someone succeeds to line up their sign vertically or horizontally without an interruption, that player wins. With our help you can make a multiplayer browser game, which you can play with a friend.

here I have used HTML, CSS and JavaScript to create a basic website of this game. 
The code is quite simple and easy to understand. 
The functions are user friendly and interactive.